vim-plugin: Add plugin for navigating undo tree
Navigate the tree with <j>, <k>, and <CR> to revert to a given state.
Use <p> to show the diff between your _current_ state and the selected
one, instead of showing a diff of the selected change against the
previous.

The undo tree can be searched with </>, and the search results can be
navigated with <n> and <N> as usual.