vim: Add clangtidy fixer
Some clang-tidy checks provide fixes. But it seems like not all checks
that provide fixes are used when I run the fixers. Further, caution is
necessary, because there are some fixes I don't want to apply.

Also add the jedi-language-server because a certain work project doesn't
work with pyls for some unknown reason.