vim: Make selected comments readable even in inactive panes
I use Tilix as my terminal emulator, which has the ability to dim
inactive panes. However, when the cursor is on a comment, or a commented
is visually selected and the pane is dimmed, the comment has the same
color as the CursorLine and Visual selection.

So lighten the comment color slightly to make them more readable.

However, Tilix only dims the 16 base colors, so comments won't be dimmed
along with the rest of the pane contents. I think this is acceptable.