vim: Fix spellfile sorting altscreen bug
By default, even if you use "silent!", "execute" will move the prompt to
the bottom of the screen, because if you didn't hide the shell output,
Vim wants it to display at the bottom of the screen, because that's
where you typed the command.

However, this is jarring if there was no external command run, or if
there was no output. Jarring enough, that I disabled sorting the
spellfile because of it.

But if you use the vim-altscreen plugin, it configures t_ti and t_te
such that external commands are output on the altscreen [1]. However,
that means you have to wait to sort the spellfile until the altscreen
has been configured, so I use an autocmd.

Closes #57

[1] https://vi.stackexchange.com/questions/24645/difference-between-cmd-and-w-cmd-regarding-alternate-screen