bash: Reorder script sourcing to fix PS1 bug
Problem:

    nots@chaos ~/.config/dotfiles (ag/work *>) $ bash
    nots@chaos ~/.config/dotfiles $ bash -l
    nots@chaos ~/.config/dotfiles (ag/work *>) $

Starting a new shell would not display the __git_ps1 unless started as a
login shell.

The __git_ps1 is added to my PS1 by:

    # Test whether the given function exists.
    function function_exists() {
        declare -f -F "$1" >/dev/null
        return $?
    }

    # Decorate PS1 with git branch, rebase, cherry-pick state.
    # Further customized by $DOTFILES_DIR/lib/06-environment.sh
    if function_exists __git_ps1; then
        PS1="${PS1}\[${BLUE}\]\$(__git_ps1)\[${RESET}\]"
    fi

If you start the subshell with "bash -x", and search for where the PS1
is set, you can see that the "function_exists __git_ps1" test failed.

This is because the Bash completions that provide __git_ps1 were not
sourced yet. The reason they _are_ sourced with a "bash -l" login shell,
is because they come from the _system_ /etc/profile, which in turn,
sources /etc/profile.d/bash_completion.sh.