vim: Do not use two spaces when joining sentences
When joining two lines where the first line is the end of a sentence,
Vim will use two spaces to separate the sentences.

This is a holdout from using typewriters, and is, quite frankly, wrong.
In my experience, two-spacers either do so out of habit, or because they
need some reason to feel special.

If you're looking for typographic goodness, use TeX (which doesn't
double-space; it's actually only _slightly_ wider). Otherwise, get off
my lawn.