clip: Detect stdin/stdout redirections to set operation
    $ echo "foo" | clip
    $ clip
    foo
    $ clip <<EOF
    > bar
    > EOF
    $ clip
    bar
    $ clip | cat
    bar
    $ clip >/dev/tty
    bar
    $ clip >/tmp/clip.txt
    $ cat /tmp/clip.txt
    bar

A known problem (and I think it's an awkward behavior of wl-copy) is
that if you're copying AND pasting, the operation works, but then hangs.

    $ echo "foo" | clip | cat
    bar
    <hangs>

I also tried an alternative implementation where "echo foo | clip | cat"
would output "foo" (just pass through anything given) with a named pipe

    pipe="$(mktemp --dry-run /tmp/clip-XXXXXXXX)"
    mkfifo "$pipe"
    tee "$pipe" | wl-copy
    cat "$pipe"
    rm "$pipe"

but this hung too (interestingly, on the wl-copy invocation?!). I'm not
sure what's going on, but it only happens on chained pipes; the
following works just fine

    $ echo "foo" | clip >/tmp/clip.txt

I'm not sure how useful a feature dropping 'clip' into the middle of a
pipeline is, and if it _is_ useful, I'm not sure what the desired
behavior should be:

1. Should it pass through the input?
2. Should it output the clipboard's contents before copying new
   contents?