vim: Do not always open the quickfix window after :make
See also: https://github.com/tpope/vim-dispatch/issues/ 310

makeprg is a per-buffer setting, because it depends on the compiler,
which depends on your filetype. So vim-dispatch has to save and restore
the makeprg. It's setup to do so, but it only works if nothing else
switches buffers (like having a QuickFixCmdPost that switches focus
_while_ vim-dispatch is cleaning up after your makeprg invocation).

Since 1, I don't use :make in my normal workflows, and 2, :Make from
vim-dispatch is better (and "just works"), then remove the offending
autocmd.