bash: Enable using tmux installed in ~/.local
There's a few new features and bug fixes in tmux 3.3 and 3.4 that
resulted in me wanting to build tmux from source. But that means that it
must be in the $PATH before trying to run it.

Additionally, don't exit, because I had misunderstood what happens. I
thought that the bashrc would get sourced twice. Tmux is interactive,
and starts a new shell, so exiting would only exit the whole terminal
emulator when the last tmux window is exited.