Add script to send notification to my phone after command finishes
Note that for at least the current versions of Android, KDEConnect (the
phone app) and GSConnect (the GNOME extension) this requires the phone
be unlocked to use GSConnect.

At home, that's not a big problem, because my home is a Trusted Place.

Also note that you need to enable sending notifications from Linux to
Android, and that you probably don't want notifications from Spotify or
Teams, which you need to turn off from the GNOME extension settings.