uhist: Remove extraneous commands with subcommands
The purpose of the multi-command logic is so that commonly used things
like git and docker don't inflate the histogram, since they are both
very common, and have a wide range of uses. Commands should only be
added to this list if I use them frequently enough they'd still be in
the top ~50 commands if they were broken out into subcommands.

Additionally, the COMMAND_WRAPPERS didn't work, because I was testing
the whole command argv list against the COMMAND_WRAPPERS, instead of
just the first element :(