teams: Try fixing status indicator
I've gotten reports that I always show as online. I can't tell if this
option should be true or false, and I'm wondering if it only works with
X11 ... (discord on Wayland has the same problem).

I suspect that neither 'true' nor 'false' resolves the always-online
issue. I expect it might be a Wayland thing?

The --help output says this:

    --awayOnSystemIdle                Sets the user status as away when
                                      system goes idle
                                             [boolean] [default: false]

which makes me think that 'true' is *expected* to set the status to Away
when the system is idle, so that's what I'll set the config to, and then
maybe it'll work eventually.